Refactor: move CandidateState to candidate_state.rs
